rename hwloc module to topology module
