+++ device-api/Cargo.toml
@@ -8,7 +8,7 @@ license = { workspace = true }
 homepage = { workspace = true }
 repository = { workspace = true }
 readme = { workspace = true }
-build = "src/hwloc/build.rs"
+build = "src/topology/build.rs"
 
 [features]
 blocking = [] # Enable blocking APIs
