+++ device-api/src/lib.rs
@@ -81,13 +81,13 @@ mod config;
 mod devfs;
 mod device;
 mod error;
-pub mod hwloc;
 pub mod hwmon;
 mod list;
 pub mod perf_regs;
 pub mod proc;
 mod status;
 mod sysfs;
+pub mod topology;
 
 /// List all Furiosa NPU devices in the system.
 ///
